[API] Introduce sessionPublicationService
